fix test, commonize code, split out actors
