+++ nexus/src/db/datastore.rs
@@ -89,27 +89,6 @@ impl DataStore {
         DataStore { pool }
     }
 
-    /// Constructs a DataStore for use in test suites that has preloaded the
-    /// built-in users, roles, and role assignments that are needed for basic
-    /// operation
-    #[cfg(test)]
-    pub async fn new_with_builtins(
-        logctx: &dropshot::test_util::LogContext,
-        pool: Arc<db::Pool>,
-    ) -> Arc<DataStore> {
-        let datastore = Arc::new(DataStore::new(pool));
-        let opctx = OpContext::for_background(
-            logctx.log.new(o!()),
-            Arc::new(authz::Authz::new()),
-            authn::Context::internal_db_init(),
-            Arc::clone(&datastore),
-        );
-        datastore.load_builtin_users(&opctx).await.unwrap();
-        datastore.load_builtin_roles(&opctx).await.unwrap();
-        datastore.load_builtin_role_asgns(&opctx).await.unwrap();
-        datastore
-    }
-
     // TODO-security This should be deprecated in favor of pool_authorized(),
     // which gives us the chance to do a minimal security check before hitting
     // the database.  Eventually, this function should only be used for doing
@@ -2354,33 +2333,55 @@ impl DataStore {
     }
 }
 
+/// Constructs a DataStore for use in test suites that has preloaded the
+/// built-in users, roles, and role assignments that are needed for basic
+/// operation
+#[cfg(test)]
+pub async fn datastore_test(
+    logctx: &dropshot::test_util::LogContext,
+    db: &omicron_test_utils::dev::db::CockroachInstance,
+) -> (OpContext, Arc<DataStore>) {
+    let cfg = db::Config { url: db.pg_config().clone() };
+    let pool = Arc::new(db::Pool::new(&cfg));
+    let datastore = Arc::new(DataStore::new(pool));
+
+    // Create an OpContext with the credentials of "db-init" just for the
+    // purpose of loading the built-in users, roles, and assignments.
+    let opctx = OpContext::for_background(
+        logctx.log.new(o!()),
+        Arc::new(authz::Authz::new()),
+        authn::Context::internal_db_init(),
+        Arc::clone(&datastore),
+    );
+    datastore.load_builtin_users(&opctx).await.unwrap();
+    datastore.load_builtin_roles(&opctx).await.unwrap();
+    datastore.load_builtin_role_asgns(&opctx).await.unwrap();
+
+    // Create an OpContext with the credentials of "test-privileged" for general
+    // testing.
+    let opctx =
+        OpContext::for_unit_tests(logctx.log.new(o!()), Arc::clone(&datastore));
+
+    (opctx, datastore)
+}
+
 #[cfg(test)]
 mod test {
+    use super::datastore_test;
     use crate::authz;
-    use crate::context::OpContext;
-    use crate::db;
     use crate::db::identity::Resource;
     use crate::db::model::{ConsoleSession, Organization, Project};
-    use crate::db::DataStore;
     use crate::external_api::params;
     use chrono::{Duration, Utc};
     use omicron_common::api::external::{Error, IdentityMetadataCreateParams};
     use omicron_test_utils::dev;
-    use std::sync::Arc;
     use uuid::Uuid;
 
     #[tokio::test]
     async fn test_project_creation() {
         let logctx = dev::test_setup_log("test_project_creation");
         let mut db = dev::test_setup_database(&logctx.log).await;
-        let cfg = db::Config { url: db.pg_config().clone() };
-        let pool = Arc::new(db::Pool::new(&cfg));
-        let datastore =
-            Arc::new(DataStore::new_with_builtins(&logctx, pool).await);
-        let opctx = OpContext::for_unit_tests(
-            logctx.log.new(o!()),
-            Arc::clone(&datastore),
-        );
+        let (opctx, datastore) = datastore_test(&logctx, &db).await;
         let organization = Organization::new(params::OrganizationCreate {
             identity: IdentityMetadataCreateParams {
                 name: "org".parse().unwrap(),
@@ -2405,7 +2406,7 @@ mod test {
             datastore.organization_fetch(organization.name()).await.unwrap();
         assert!(organization_after_project_create.rcgen > organization.rcgen);
 
-        let _ = db.cleanup().await;
+        db.cleanup().await.unwrap();
         logctx.cleanup_successful();
     }
 
@@ -2413,10 +2414,7 @@ mod test {
     async fn test_session_methods() {
         let logctx = dev::test_setup_log("test_session_methods");
         let mut db = dev::test_setup_database(&logctx.log).await;
-        let cfg = db::Config { url: db.pg_config().clone() };
-        let pool = Arc::new(db::Pool::new(&cfg));
-        let datastore =
-            Arc::new(DataStore::new_with_builtins(&logctx, pool).await);
+        let (_, datastore) = datastore_test(&logctx, &db).await;
         let token = "a_token".to_string();
         let session = ConsoleSession {
             token: token.clone(),
@@ -2462,7 +2460,7 @@ mod test {
         let delete_again = datastore.session_hard_delete(token.clone()).await;
         assert_eq!(delete_again, Ok(()));
 
-        let _ = db.cleanup().await;
+        db.cleanup().await.unwrap();
         logctx.cleanup_successful();
     }
 }
